Move CheckRepeatedModifier, remove MockedInput from the editor tests
