+++ src/modules/keyboardmanager/KeyboardManagerEditorTest/KeyboardManagerEditorTest.vcxproj
@@ -40,14 +40,12 @@
   <ItemGroup>
     <ClCompile Include="BufferValidationTests.cpp" />
     <ClCompile Include="LoadingAndSavingRemappingTests.cpp" />
-    <ClCompile Include="MockedInput.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(CIBuild)'!='true'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="EditorHelpersTests.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="MockedInput.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
