--- src/modules/keyboardmanager/KeyboardManagerEngineTest/HelperTests.cpp
@@ -1,59 +1,0 @@
-#include "pch.h"
-#include "CppUnitTest.h"
-#include <keyboardmanager/common/ErrorTypes.h>
-#include <keyboardmanager/common/Helpers.h>
-#include "TestHelpers.h"
-#include <common/interop/keyboard_layout.h>
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-namespace KeyboardManagerCommonTests
-{
-    // Tests for methods in the Helpers namespace
-    TEST_CLASS (HelpersTests)
-    {
-    public:
-        TEST_METHOD_INITIALIZE(InitializeTestEnv)
-        {
-        }
-
-        // Test if the CheckRepeatedModifier method returns true on passing vector with same modifier repeated
-        TEST_METHOD (CheckRepeatedModifier_ShouldReturnTrue_OnPassingSameModifierRepeated)
-        {
-            // Arrange
-            std::vector<int32_t> keys = { VK_CONTROL, VK_CONTROL, 0x41 };
-
-            // Act
-            bool result = Helpers::CheckRepeatedModifier(keys, VK_CONTROL);
-
-            // Assert
-            Assert::IsTrue(result);
-        }
-
-        // Test if the CheckRepeatedModifier method returns true on passing vector with conflicting modifier repeated
-        TEST_METHOD (CheckRepeatedModifier_ShouldReturnTrue_OnPassingConflictingModifierRepeated)
-        {
-            // Arrange
-            std::vector<int32_t> keys = { VK_CONTROL, VK_LCONTROL, 0x41 };
-
-            // Act
-            bool result = Helpers::CheckRepeatedModifier(keys, VK_LCONTROL);
-
-            // Assert
-            Assert::IsTrue(result);
-        }
-
-        // Test if the CheckRepeatedModifier method returns false on passing vector with different modifiers
-        TEST_METHOD (CheckRepeatedModifier_ShouldReturnFalse_OnPassingDifferentModifiers)
-        {
-            // Arrange
-            std::vector<int32_t> keys = { VK_CONTROL, VK_SHIFT, 0x41 };
-
-            // Act
-            bool result = Helpers::CheckRepeatedModifier(keys, VK_SHIFT);
-
-            // Assert
-            Assert::IsFalse(result);
-        }
-    };
-}
